requirements.txt
anyio==3.6.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dependency-injector==4.40.0
docker-pycreds==0.4.0
docopt==0.6.2
fastapi==0.86.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.1
httptools==0.5.0
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
Mako==1.2.3
MarkupSafe==2.1.1
packaging==21.3
pipreqs==0.4.11
pluggy==1.0.0
postgres==4.0
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
pytest-docker==1.0.1
pytest-mock==3.10.0
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.43
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
watchfiles==0.18.1
websocket-client==1.4.2
websockets==10.4
yarg==0.1.9
